Use spl_object_hash() for Notifiable IDs
This replaces the dependency on the getKey() method, which may not be available if the Notifiable is not an Eloquent model.

Since a Notifiable can be extremely flexible and doesn’t follow a specific interface, spl_object_hash() works well as an identifier for an object instance in a pinch.