transChoice instead of trans for error messages
For some languages, messages need to be formatted differently based on the attribute word gender (masculine or feminine). And based on this update, developers can define the `attribute_gender` under the `resources/lang/{LANG_CODE}/validation.php` by using either `m` value for "masculine" or `f` for "feminine". On the other hand, the message could be written in the two choices (0 for masculine and 1 is for feminine):

`'required' => ':attribute مطلوب.|:attribute مطلوبة.'`

In either case, if there's no definition for the "attribute_gender" or there weren't choices for the message itself, it will fallback to the default value.

Please, take a look at this snippet for further details: 
https://gist.github.com/majebry/e598c335777fdbb0c671a987b998cdfb