[5.7] Switch merging order on HasManyThrough::shouldSelect to prevent overriding same foreign key
This commit will fix an issue that can be reproduice when we try to access a model w/ the HasManyThrough relationship, the resulting model(s) if they happen to contain the same column name (in the test `Project` contains also owner_id).

When trying to access the user's owned project through Team, the result will be the right Project but if you try to access the owner again, the foreign key owner_id will be override by the owner_id used in the HAsManyThrough relationship causing a weird but where it's impossible to access the right owner again unless you refresh the model.