Use default INSERT INTO syntax for multiple records with SQLite
This bumps the version of SQLite dependency to 3.7.11, released
March 2012.

Using a default syntax for the INSERT INTO clause fixes an issue
described in the bug #25262, but only when used with SQLite 3.8.8
(released January 2015) or newer.

The DEFAULT VALUES case needs to be still handled though, just like
in the PostgresGrammar.

Closes #25262