Illuminate\View: Normalize first level data keys passed upon view creation
Similar to what VueJS does when passing props using kebab-case whch are automatically converted
into accessible camelCase variables, I wanted to have a go to see if it would be possible to
avoid "loosing" a variable when passing it via @include() or @component() data array.
I know it's not properly lost as it's always accessible with `$__data['kebab-case']` but I wanted
to see if I could make it easier for developers to retrieve it. This should be of course only
applied to the first level keys of the data array.

* Implementing a new normalizeDataKeys() method on ViewFactory using official regex to identify if the key is legal or not.
* Make sure this is called on make() and file() methods only for data passed by the developer, we don't want this to be applied to `$__env` for example.
* Covering the functionality with unit tests.
* Make sure tests/View tests are green.

* Item