Access the $manyMethods property through static::$manyMethods
This allows adding additional $manyMethods when extending the model class
and still receive a correct guess for the relationship name.

Removed the name of guessBelongsToManyRelation from the list but still consider
it when analyzing the call stack. This makes $manyMethods more slim and easier
to override.

Removed findFirstMethodThatIsntRelation from $hasMany array, since it does not
appear anywhere in the code and seems to have been forgotten there.
It is also not really a name for a many-relationship method, so if it is needed,
i think there might be a better place for it.