Cast key value to string when your related table uses string instead of int
When working with an database that was managed remote we ran into performance issues. The indexes on columns were not used. After a EXPLAIN EXTENDED we noticed this was a issue with mismatching data types.

The parent table uses INT for the primary key but the related tables uses varchar for their fields. Because this database is not ours we can not change this!

The suggested change fixes this issue. It uses the existing KeyType that is present on the Model instance.