Strip keys from array when converting paginator objects to array
At the moment when paginator objects with an item set without a standard sequence of numbers starting from 0 are being compiled to JSON the data key will be converted to an object rather and an expected array. This commit will always make sure that the data key is an array and that any keys are stripped when converting to JSON.

For this a new `compile` method was added with an option to strip the keys from the data array. This approach allows for the `toArray` to keep its current behavior and only modify the behavior when converting to JSON.