Allow using Mail::fake for mailables sent via Notifications
Currently, if an email is sent via a Notification, the MailChannel
sends the message directly via the Mailable send method.  This causes
problems when Mail::fake is activated in tests.  The send method in
`MailFake.php` expects to see a Mailable passed as the view, but when
the mailable itself calls send, it gets passed the result of the call
to `buildView` on line 153 of Mailable.php's send method. This causes
the assertion that a mail was sent to fail, because MailFake
returns when it sees the `$view` is not an instance of a Mailable.
This change ensures that the pattern implemented by Mailer and
Mailables still works correctly in a mocked environment when sent
via notifications.