Validate $line is a string to prevent htmlspecialchars being passed array
To resolve: https://github.com/laravel/framework/issues/26664

When a user attempts to translate a key that has an array for its value,
we encounter an error as this array is passed to htmlspecialchars().

Similarly, issue is also encountered when the user attempts to translate
a word that is also used for the name of a translation file.

This fix ensures that if the string to be translated has an array value,
the original string is output to the screen.

Same behaviour will also occur if the string corresponds to the name of a
translation file.