remove extra alias
the `Swift_SendmailTransport` is aliased to both `MailTransport` and `SendmailTransport`.   We can have both the 'mail' and 'sendmail' drivers build up the same class, and remove the need for the extra alias.

I think this also makes it more obvious to the reader that the 'mail' driver doesn't really exist anymore and is actually using sendmail.