Allow cache stores to handle null TTL
At the moment our Cache implementation implements the Psr\SimpleCache\CacheInterface which states that the TTL, if null, allows for either the library or the driver to set a default. At the moment it's impossible with Laravel to allow for the drivers/stores to handle defaults on their own because the method signature doesn't allows for this. By allowing it as a valid type, developers may choose to implement a cache driver which handles a default TTL for cache items. Our current cache stores are updated to handle the new null value as well to properly return false when null is passed. This is because our current cache stores don't provide a default TTL.