Fix handling null TTLs in cache repository
The way we currently handle null TTLs in our own cache repository is that we return false and don't save anything. But according to the PSR cache repository spec there should be support for handling defaults in drivers. At the moment the null TTL isn't cascaded to the cache stores so there's no way for them to implement a default. This commit allows for the TTL to be passed to the cache stores so they may set their default TTL is they have done so.

It also still makes sure that for any TTL below or equal to 0 nothing will be set.

No tests were adjusted (just a really minor adjustment to a Redis test) and more tests have been added to verify the new behavior.

Also fixed some really small minor things here and there.