Email notification view: pass subcopy as a slot
With the current implementation, the notification subcopy skips over the mail::message component segment for it and is inserted into mail::layout via $slot instead of as $subcopy with $subcopy being empty. This becomes an issue if the $subcopy in mail::layout is moved anywhere besides directly after $slot, as it will not be inserted into the correct location.

The notification subcopy should instead be passed as a slot to the mail::message component to fill the $subcopy there, which then passes the subcopy component as a slot to mail::layout for final insertion as $subcopy instead of $slot.