Resize the redis queue notifications list
If a worker crashes after it's popped a notification but before it's
popped the job the next worker that attempts to pop the job will block
waiting for the notification unnecessarily.  To prevent this we can
check that all ready jobs have a notification and if they don't create
any missing notifications.

Similarly if a worker does not pop the notification for some reason
(possible if we are mixing workers from different versions) the notify
list can grow too long.  To prevent this we can trim the notification
list to ensure the # of notifications matches the # of ready jobs.