do not JSON encode strings
if you cast a Model attritube to 'array', 'json', 'object', or 'collection', they are stored as JSON in the database.  To set them on the model, you **must** pass either an array, object, or Collection. Currently you **cannot** pass a properly formatted JSON string that would be decoded into one of the types.  This PR allows you set a property to  a JSON string, and to skip the encoding, so when it is decoded it comes out properly.