Add support for schemas in SqlServer grammar
This merge requests adds checks for the schema name in the `compileTableExists()` and
`compileColumnListing()` functions.
The available schemas can be defined in the database configuration.
This is all consistent with the PostgresGrammar and PostgresBuilder.

The default schema is gathered from the database by using
```
select SCHEMA_NAME()
```
Which is available since SQL Server 2008.
See https://docs.microsoft.com/en-us/sql/t-sql/functions/schema-name-transact-sql?view=sql-server-2017