Tweak path helper functions to use direct function calls
Some path helper functions used the direct function calls and some used the Service Container binds and appended path suffixes themselves, which is duplicated code from the actual functions in Application.php. This change actually makes it so that the helper functions call the function directly from Applciation.php and pass the path suffix to the function which eventually deals with appending the suffix.