Underscore before numbers, too
When trying to fetch database columns in Eloquent ORM models with mutated attributes, it uses `Str::snake($match)`. However, when trying to use a mutator to access `billing_address_line_2` when the row is changed into JSON, will cause Eloquent ORM to assume that we're looking for the attribute `billing_address_line2` which doesn't exist. Snake casing in other situation always has an `_` before the numbers, just as Laravel Spark did with the DB column name.