[5.8] Added a Blade directory `@includeAll` for dynamic including of views
The function allows you to automatically include all template files from the specified folder.

For example:
```
// resources:
resource
|-views
| - index.blade.php
| - foo
| | - bar
| - | - q1.blade.php
| - | - q2.blade.php
| - | - q3.blade.php
```

```
// resources/views/index.blade.php
// include all files from `resources/views/foo/bar` folder
@includeAll('foo.bar')
```

Where does it apply?

In several projects, we often change SEO scripts, so it is not a good idea to prescribe each file manually.
This directive allows you to specify a folder from which all views files will be automatically downloaded.

![2019-06-07_16-04-25](https://user-images.githubusercontent.com/10347617/59106029-036d8e80-893e-11e9-858e-782d2cecd5a6.png)