compile `@endcomponentFirst`
while the `@endcomponent` directive is perfectly valid to end a `@componentFirst`, the IDEs struggle to match the 2 up resulting in a syntax error, which I'm sure people will be upset about.

I would prefer the plugin handle this if it can, but I'm not sure if it can, or where to change that, so I'm PR'ing here for now.