Fix bug in paginate method in Builder.php
Laravel has facility to pass columns (having clause columns define in select) in paginate method but it won't pass it to getCountForPagination() method which requires columns which can come from paginate method. Due to this, it is not able to add having clause in first count agreegator query hit to get total no of records.