Prevent a job from firing if it's been marked as deleted
During queue job processing an `Events\JobProcessing` event is fired, which
makes it possible to `$job->delete()`, however, `$job->fire()` still happens
even on the deleted job,.

This adds a check to the processing of the job to prevent that, and fires an
`Events\JobDeleted` event that can be listened for.

Also adds output to the `WorkCommand` to allow seeing that a job was deleted
before processing completed.

Tests included.