Queue Worker: Rename method to reflect reality
Since #29224 was merged, this class property really is an instance
of the factory contract, not necessarily a full manager.

This commit tries to prevent confusion by fixing all typehints,
removing the namespace alias and renaming the accessor methods.

These methods were not used anywhere in the framework, so this
should be a relatively minor BC break.

Refs #29224, #29395.