Proposal for more clear error instead of
`Failed to clear cache. Make sure you have the appropriate permissions` when really that folder just needs to be made. 
In my example it was deleted during a build and from there 
```
php artisan cache:clear
```
Gave an error which is not correct
"Failed to clear cache. Make sure you have the appropriate permissions"