Replace string commands with class names
This is a follow-up for https://github.com/laravel/framework/pull/29624. By replacing these commands with their class names it's more clear which commands actually are called. When changes need to be made to any command signatures it can be done so in one place. When using an IDE it also allows for easy click through to these classes to inspect them and make use of any refactoring tools when change is needed.