Create storage fakes with custom configuration
Currently when a storage disk is faked, we are creating a local disk with the default configuration.

If the developers want to use a custom configuration or reuse the configuration they have in their `filesystems.php` file, there is no easy way to achieve this.

This commit allows us to pass optional configuration as a second parameter:

```php
Storage::fake('documents', ['visibility' => 'private']);
```

This change is totally backwars-compatible, but in the future it would be convenient to fetch the actual configuration of the given disk (maybe in Laravel 7.x?).