Ensure Builder::findOrFail with Arrayable throws ModelNotFoundException
I found a regression in #7048 and #19019, where the Eloquent Builder will not throw a `ModelNotFoundException` when an `Arrayable` is passed to `findOrFail`.

In this #7048, we are checking if the `$ids` is an array and we count the results against the number of ids.

But since #19019, the `find` method also accepts an `Arrayable` for the ids.

So if an `Arrayable` is passed, the check is skipped and the method returns the results.

To fix this, we are first checking if the `$ids` is an  `Arrayable` and we convert the ids to an array before checking the results.