Dispatch event when a scheduled command was filtered from running
If a scheduled command is run with a truth constraint, it is not simple
to check whether the task didn't run at all, or if it was filtered
running from a known system constraint.

Firing an event when the task is intentionally skipped allows monitoring
over the task executing as scheduled, based on the truth constraint.