Extends tap helper function to continually return instances.
The tap helper function returns the same instance it is created with but
with this change it will perpetually return the same instance after each
method call.

Before this change the following would work:
class A
{
    private $b;

    public function __construct($b = 0)
    {
        $this->b = $b;
    }

    public function increment()
    {
        $this->b += 1;
    }
}
tap(new A())
   ->increment()  // $b is 1
   ->increment(); // $b is 2

But an error would result is further calls:
tap(new A())
   ->increment()  // $b is 1
   ->increment()  // $b is 2
   ->increment(); // PHP Error - Call to a member function increment on null

After this change the instance is continually returned, allowing
perpetual calls. Now the following will work:
tap(new A())
   ->increment()  // $b is 1
   ->increment()  // $b is 2
   ->increment()  // $b is 3
   ->increment()  // $b is 4
   ->increment(); // $b is 5