Update Mix.php to not fallback to localhost:8080
Right now when the public/hot folder is present (and empty), it has a fallback to //localhost:8080, which makes no sense for environments other than a local one (like php artisan serve or Homestead). We're currently testing our new automatic deployments which ensures folders are present and symlinked. When creating the hot folder, all asset URLs (created with asset(mix())) broke, which is how we discovered this issue.

This proposed change simply creates the HtmlString from the path, which already prepends the site's URL, including port, if present. I've tested with both a domain and localhost with port and it works for both.