Auth guards setUser $user and actingAs $user can be null
Allows for actingAs on testing to be null. This is helpful when testing applications where most routes are auth protected and you have actingAs a test user in the base TestCase but you want to test certain routes as guest. You can now simply use $this->actingAs(null) or even $this->actingAs(null, 'api').