Change default date serialization on Models
This change updates the default serialization for toArray and toJson for dates. It makes uses of Carbon's default serialization to an ISO 8601 string which besides the previous values also contains fractions and the timezone. This is especially handy for people who are building API's.

Implements https://github.com/laravel/ideas/issues/1940