Test console hyperlinks with ANSI and non-ANSI output
This test asserts that console hyperlinks are rendered correctly with ANSI output enabled and disabled.

For terminals that support ANSI output and can handle hyperlinks, the `Symfony\Component\Console\Formatter\OutputFormatterStyle` will render the hyperlinks as expected. For example, this will be the case when running `php artisan foo:hyperlink` in the terminal.

When the output of the artisan command is appended elsewhere, e.g. `php artisan foo:hyperlink >> foo.log`, Symfony will know that this output is not *decorated*, and will simply render the link text. This will also be the case when storing output from the command scheduler: `$schedule->command('foo:hyperlink')->storeOutput()`.

When PHPUnit is running, the artisan binary is *not* actually executed, but instead the console Kernel calls the command using a `BufferedOutput` instance (mocked or not). In that case the output is not decorated, so it will render the plain text and not the hyperlink. This applies for the other output commands as well, like `$this->info()`.

In order to simulate a decorated output, we will pass the `--ansi` flag to the `HyperinkCommandStub` and mark the `BufferedOutput` as decorated. This will force the `OutputFormatterStyle` to actually render the hyperlink, so we can perform an assertion on this.