Specify dynamic attribute casts at query-time
This pull request allows additional (or replacement) attribute casts to be specified on the Eloquent query builder, such that they apply only to the results of that query.

The primary use case for this functionality is to apply casts to columns resulting from subqueries or joins, that are not usually present on the model.

While adding the aliases for those columns to the `$casts` array on the model does work, I feel that polluting the model with columns that are simply not present most of the time is not ideal, especially when the actual definition of those column names are likely to be spread across many different files.

Consider the following query selecting `User`s along with the date of their most recent `Post`:

```php
User
	::select([
		'users.*',
		'last_posted_at' => Post::selectRaw('MAX(created_at)')
			->whereColumn('user_id', 'users.id')
	])
	->get();
```

While `User`’s own timestamps will be cast to Carbon instances, `last_posted_at` will simply be a string. Currently, the only way to cast this would be to add `last_posted_at` to either `$casts` or `$dates` on `User`, despite the column only being present for this one specific query.

This PR allows the following:

```php
User
	::select([
		'users.*',
		'last_posted_at' => Post::selectRaw('MAX(created_at)')
			->whereColumn('user_id', 'users.id')
	])
	->withCasts(['last_posted_at' => 'date'])
	->get();
```

where the array of casts will be merged into any existing casts specified on the model.

Naming wise I’m not sure if “withCasts” is the best option; I’ve just gone along with the likes of “withCount”.