Custom Migration Table with Prefix Error
If you need to use a prefix to access the migrations table (ie. SQL Server), the migration fails to find the table in the schema.

The schema just stores the table name, so it doesn't think the migrations table exists.

Steps to reproduce:
In `config/database.php`:
Change `'migrations' => 'migrations',` to `'migrations' => '[Database/Schema Name].migrations',`

Try to run migrations or just do: `php artisan migrate:status`

This (terrible) fix, is to highlight that the table name is not directly found in the schema when using a schema prefix.

My SQL Server is not using `dbo` as the table schema prefix.