Introduce BeforeQueryExecuted event
This is a follow up of b8ecedacce80ccfe5c2f44c51c35954d836df413 that
introduced StatementPrepared event.

However, it does not provide an ease way to modify the raw SQL query
before it's prepared and executed by the database server. Further more,
only select queries trigger this event.

This PR introduces BeforeQueryExecuted event that is provided with a raw
SQL query and connection instance. SQL query string is passed as a
reference, so it can be modified by whatever binds to this event.

Use? A comment can be prepended to the query to ease debugging of
performance issues as these comments will show up in MySQL's SHOW
PROCESSLIST and in slow query log.

An example of a modified query:

/* UserQuizResultsApiController.php:96 POST /papi/v2/quiz_results */ insert into `user_quiz_results` (`created_at`, `quiz_answer_id`, `quiz_question_id`, `user_id`) values (?, ?, ?, ?)

We will publish a composer package providing the above SQL commenting
feature for Laravel shortly.