Generalize Filesystem hashing while preserving backwards compatibility with Laravel 5
The Filesystem::hash() function only allows for md5 hashes, which is
prone to collision attacks.  The PHP builtin function hash_file() allows
for more algorithms.  By default we select the md5 algorithm because the
users of this method likely can't afford to have all their hashes be
invalid all of the sudden.

In the future it will be beneficial to select a different hashing method
as the internet is moving away from md5 and to more modern hashing
standards.

This comment on php.net:
https://www.php.net/manual/en/function.hash.php#89574 outlines all of
the methods allowed by the $algo parameter and we must consider adding
it/linking it in the docs or in the docblock.  For now, I have left this
out because there is no rule in the styleguide to link to this kind of
user-contributed documentation.

Signed-off-by: Rudolf R Ford <60923682+r32rf@users.noreply.github.com>