Only define the query option when query parameters were specified
This allows multiple query parameters with the same name to be used. This is currently only possible by specifying the query parameters in the URL, rather than using the query option. If the consumer does not provide query parameters via the options, it assumed there are no additional query parameters to be passed to Guzzle.