Allow URL-encoded special characters in basic auth part of URLs
Special characters in HTTP Basic Auth passwords in an URL need to be url-encoded.
Example: foo@bar becomes foo%40bar, in an URL: http://user:foo%40bar@example.org

The UrlValidator did not allow percent signs in username and password, and this is changed now.

Related bugreport and fix for symfony:
- https://github.com/symfony/symfony/issues/36285
- https://github.com/symfony/symfony/pull/36286