Disable transaction events in DatabaseTransactions
This follows commit deafaa70e5bb9e86e30585a50e7872814a7af53e
which already added this for RefreshDatabase.

It's for the exact same reason as outlined in https://github.com/laravel/framework/pull/23832 :
> This is motivated by the fact that in tests, when using the RefreshDatabase trait, transactions events are dispatched and may interfer with the application testing itself when it relies on events like these ( for further info, see https://github.com/laravel/ideas/issues/1094 ).

As to not have custom handling of transaction events interfere with the
test suite itself, this simply disables this.

Again, the same change was added to `RefreshDatabase` and is still
present; this PR merely keeps the behaviour of them in sync in regards
to the handling of events in transactions.