Update Http Facade attach docblock to be accurate
Currently, the Http client facade has a typo that says the `$headers` parameter in the `attach` function is necessary and doesn't have a default. In the actual code (and in the documentation) it is clear that this parameter is not actually necessary and has a default value of an empty array.

This typo causes confusion in IDEs like PHPStorm, where if you leave out the `$headers` parameter it complains that you are using the function incorrectly, even though during runtime everything works fine.