Allow app config to override package aliases
If you include two packages with the same alias defined, the rule will be that the last one will override the others (PackageManifest::config ->Collection::flatMap -> Arr::collapse). Which may be tricky to determine. Allowing the app to override the definition may be a good way to fix that kind of conflict. However that change can modify actual behavior where package overrides default aliases.