Convert EloquentCollection to base if not solely filled with models
Some methods, such as `map()` or `concat()` may introduce non-model items
into an EloquentCollection. Now, this is checked not only for `map()` but
also for other methods.

Other methods can be known ahead of time to return a base Collection, in
those case the check is omitted and the call is simply passed to base.

This achieves a few things:
- Prevent bugs by nature of having more accurate return type hints
- Ensure correct behaviour for code that has special treatment for EloquentCollection
- Aid static analysis tools in flow control