Convert EloquentCollection to base if not solely filled with models
Some collection methods, may introduce non-model items
into an EloquentCollection. This was previously only checked for `map()`.
This PR extends this check to include other such methods.

Another group of methods can be known ahead of time to return a base Collection, in
those cases the check is omitted, the collection is converted `toBase()`
and the call is passed along.

This achieves a few things:
- Prevent bugs by having more accurate return type hints
- Ensure correct behaviour for code that has special treatment for EloquentCollections
- Aid static analysis tools