Make valid method names for abilities with spaces
If an ability has a space in its name, such as "edit post", the current Gate implementation will check if there's a method named "edit post" in the model's policy. That method obviously doesn't exist because it's invalid PHP. I changed the `formatAbilityToMethod()` method to check for 'editPost' in these situations instead. I think that's intuitive behavior - more intuitive than renaming your ability to "edit-post" as you are now supposed to do.