Slight optimization to mergeConfigFrom
For cold starts and lumen, this avoids 2 magic method calls to ArrayInterface
and one make call which itself has a bit of logic. This should not
affect laravel with config caching but config is in a hot path for requests in
lumen it is a small change that could add up.