Fatal error handling memory limit increase
When an `Out of Memory` Fatal error occurs the `$reservedMemory` allocated is not enough when the memory limit is hit due to a large DB query (both Sentry and BugSnag were unable to report many of our out of memory errors until we increased this value). If you don't think the default should be updated, then it should at least be user-configurable as it's very important that memory limit errors not go unnoticed/unreported.  Thanks!