Add a test.
The test just checks that the SupportTestTraitRecursiveB trait uses
SupportTestTraitRecursiveA. The two traits use each other, so if
trait_uses_recursive() can't handle traits that point at each other there
will be a "Maximum function nesting level" error when running the test.

Traits have to be in separate files because we get a PHP "Trait not found"
error when declaring the 1st of the two traits. It seems that we can't
use a trait that's not declared yet. Autoloading gets around this.