Update Factory.php
No longer creating a new factory when editing properties throught their respective methods.

This is causes issues when trying to implement an OOP style structure in your factories.
E.g. when an afterCreate Closure (registered in the `configure()` method) gets called, the $this->faker property is null because `$this->faker = $this->withFaker()` (inside `getRawAttributes()`) got called on a different instance of the factory.
Furthermore when setting (custom)properties to the factory within e.g. `definition()` those properties are not visible within the in `configure()` registerd closures because those properties would have been set on a different instance.