Maintenance mode: Fix empty Retry-After header
When maintenance mode is enabled with a prerendered template, an empty
Retry-After header is sent when the site is accessed.

This is because if `--render` is the only option set, the Retry-After
header is always sent, regardless of whether `--retry` was set.

Let's check for the presence of a `retry` value before sending the
`Retry-After` header.