Replace `static::` calls on non-static funcs with `$this->`
Updates a few method calls that use `static::foo()` pattern instead of standard `$this->foo()` pattern on non-static methods.

In PHP 8.0, it throws a fatal error if non-static methods are called with static resolver (`::`), but [with an exception for `static::`, `parent::`, and `self::`](https://php.watch/versions/8.0/non-static-static-call-fatal-error#static-this), so this is not a change required for PHP 8.0.

However, I think this change makes the code a bit more correct using the correct `$this->` calling pattern.