Allowing httpOnly from Session Config
To allow `httpOnly` value to be use from `config('session.http_only')`. 

Currently it's been hardcoded as `false`. 

For some organisation like currently am working on, it's required to set it to true, unless there's necessity to have set to false.

Hence, the option for `httpOnly` need to be based on session config rather than hardcoded it.