[8.x] Fix validating image/jpeg images after Symfony/Mime update
Symfony/Mime v5.20 changed the returned extension for an image with image/jpeg mime type: https://github.com/symfony/mime/commit/aa1d922a7c32aedc10f6647ded250535ee3c6346

This breaks laravel image validation, as it relied on a return value of `.jpeg`. This commit changes `.jpeg` to `.jpg`, which fixes the validation. I coudn't find any tests for this part of the framework, but I'm not very familiar with the framework tests, so this might be an oversight on my part.