Fix issue with special $user search path variable not being resolved
For any scenario in which PostgreSQL's internal "information_schema" is queried using a schema/schemas derived from the search_path configured on the connection, it is necessary to resolve PostgreSQL's special "$user" variable to the actual username set on the connection.

While this resolution is performed correctly in the related parseSchemaAndTable() method, it was not being performed here, prior to this change.